+++ CHANGELOG.md
@@ -5,6 +5,37 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+
+## [1.6.0] - 2022-07-27
+
+### Added
+- Pipeline functionality analysis working with MMSA Shelby buried pipelines [#175](https://github.com/IN-CORE/pyincore/issues/175)
+- Electric power network functionality analysis working with MMSA Shelby [#178](https://github.com/IN-CORE/pyincore/issues/178)
+- Water facility network functionality analysis working with MMSA Shelby [#103](https://github.com/IN-CORE/pyincore/issues/103)
+
+### Changed
+- Network utils refactored to use network dataset [#149](https://github.com/IN-CORE/pyincore/issues/149)
+- EPF restoration uses damage to compute discretized functionality [#169](https://github.com/IN-CORE/pyincore/issues/169)
+- Water facility restoration uses damage to compute discretized functionality [#170](https://github.com/IN-CORE/pyincore/issues/170)
+- Household-level housing sequential recovery uses social vulnerability analysis result [#168](https://github.com/IN-CORE/pyincore/issues/168)
+- Social vulnerability no longer requires year, state, county and census tract as input parameters [#152](https://github.com/IN-CORE/pyincore/pull/156)
+
+### Fixed
+- Fix data type of Census input dataset to CSV [#166](https://github.com/IN-CORE/pyincore/issues/166)
+
+## [1.5.0] - 2022-06-29
+
+### Added
+- PyPi description and README.rst [#150](https://github.com/IN-CORE/pyincore/issues/150)
+- Earthquake liquefaction to building damage analysis [#155](https://github.com/IN-CORE/pyincore/issues/155)
+- When releases are made, now push builds to pypi (or testpypi) automatically
+
+### Changed
+- Made pyincore build with legacy naming for pypi publish [#138](https://github.com/IN-CORE/pyincore/issues/138)
+- Network dataset's sub category's dataType has been changed from networkType to dataType [#145](https://github.com/IN-CORE/pyincore/issues/145)
+- Tornado EPN damage analysis uses network dataset instead of link, node, graph datasets [#147](https://github.com/IN-CORE/pyincore/issues/147)
+- Building functionality to compute functionality without power network [#143](https://github.com/IN-CORE/pyincore/issues/143)
+
 ## [1.4.1] - 2022-04-22
 
 ### Fixed
@@ -14,7 +45,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ## [1.4.0] - 2022-03-30
 
 ### Added
-- Add check to mean damage analysis to verify damage keys match inventory type and remove unsupported types [#53](https://github.com/IN-CORE/pyincore/issues/53)
+- Check to mean damage analysis to verify damage keys match inventory type and remove unsupported types [#53](https://github.com/IN-CORE/pyincore/issues/53)
 - Housing recovery model analysis [#99](https://github.com/IN-CORE/pyincore/issues/99)
 - Social vulnerability analysis [#106](https://github.com/IN-CORE/pyincore/issues/106)